requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==2.1.1
cloudinary==1.30.0
Django==4.1.4
django-cloudinary-storage==0.3.0
django-crispy-forms==1.14.0
django-ipware==4.0.2
gunicorn==20.1.0
idna==3.4
mysqlclient==2.1.1
Pillow==9.3.0
requests==2.28.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
whitenoise==6.2.0